+++ package.json
@@ -13,7 +13,7 @@
   "license": "ISC",
   "description": "",
   "dependencies": {
-    "@types/multer": "^1.4.12",
+  
     "bcrypt": "^5.1.1",
     "cloudinary": "^2.5.1",
     "cors": "^2.8.5",
@@ -32,6 +32,7 @@
     "winston": "^3.17.0"
   },
   "devDependencies": {
+    "@types/multer": "^1.4.12",
     "@types/bcrypt": "^5.0.2",
     "@types/cors": "^2.8.17",
     "@types/express": "^5.0.0",
