+++ src/middleware/upload.ts
@@ -1,0 +1,62 @@
+// import multer from 'multer';
+// import { Request } from 'express';
+
+// const storage = multer.memoryStorage();
+
+// const fileFilter = (req: Request, file: Express.Multer.File, cb: multer.FileFilterCallback) => {
+//   // Accept images only
+//   if (!file.originalname.match(/\.(jpg|jpeg|png|gif)$/i)) {
+//     return cb(new Error('Only image files are allowed!'));
+//   }
+//   cb(null, true);
+// };
+
+// export const upload = multer({
+//   storage: storage,
+//   limits: {
+//     fileSize: 5 * 1024 * 1024, // 5MB limit
+//   },
+//   fileFilter: fileFilter
+// });
+
+
+// import multer from 'multer';
+// import { Request } from 'express';
+
+// const storage = multer.memoryStorage();
+
+// const fileFilter = (req: Request, file: Express.Multer.File, cb: multer.FileFilterCallback) => {
+//   // Accept images only
+//   if (!file.originalname.match(/\.(jpg|jpeg|png|gif)$/i)) {
+//     return cb(new Error('Only image files are allowed!'));
+//   }
+//   cb(null, true);
+// };
+
+// export const upload = multer({
+//   storage: storage,
+//   limits: {
+//     fileSize: 5 * 1024 * 1024, // 5MB limit
+//   },
+//   fileFilter: fileFilter
+// });
+
+
+
+
+
+import multer from 'multer';
+
+// const storage = multer.diskStorage({
+//   destination: (req, file, cb) => {
+//     cb(null, 'uploads/');
+//   },
+//   filename: (req, file, cb) => {
+//     cb(null, `${Date.now()}-${file.originalname}`);
+//   },
+// });
+
+// const upload = multer({ storage });
+const upload = multer({ dest: 'uploads/' }); 
+
+export default upload;
