+++ src/types/express/index.d.ts
@@ -1,0 +1,10 @@
+import * as express from 'express';
+
+declare global {
+  namespace Express {
+    interface Request {
+      file?: Multer.File;
+      files?: Multer.File[];
+    }
+  }
+}
